pyproject.toml
[project]
name = "thvote-be-re"
version = "0.2.0"
description = "THVote Backend - Python refactored version"
requires-python = ">=3.12"
dependencies = [
    "alembic",
    "argon2-cffi",
    "bcrypt",
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "pydantic-settings",
    "sqlalchemy",
    "asyncpg",
    "asyncmy",
    "aiosqlite",
    "redis",
    "aioredis",
    "redis-om",
    "PyJWT",
    "python-dotenv",
    "httpx",
    "email-validator",
    "strawberry-graphql[fastapi]>=0.219.0",
    "dependency-injector>=4.41.0",
]
